+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Fix `clusterHighlights` type.
+
 ## [1.53.0] - 2021-09-02
 
 ### Changed