+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.53.1] - 2021-09-14
+
 ### Fixed
 - Fix `clusterHighlights` type.